notify & update: merge the separate TSigInfo instances
